add test to Component modify test to Directive
