+++ test/specs/files/component-i18nFiles.js
@@ -1,0 +1,5 @@
+module.exports = [
+    'src/components$prefix$name/i18n/translations.js',
+    'src/components$prefix$name/i18n/pt.js',
+    'src/components$prefix$name/i18n/en.js'
+];
