+++ test/specs/files/componentFiles.js
@@ -1,0 +1,10 @@
+module.exports = [
+    '.yo-rc.json',
+    'src/components$prefix$name$name-component-controller.js',
+    'src/components$prefix$name$name-component.js',
+    'src/components$prefix$name/index.js',
+    'src/components$prefix$name$name-component.scss',
+    'src/components$prefix$name$name-component.html',
+    'src/components$prefix$name$name-component-spec.js',
+    'src/components$prefix$name$name-component-test.js'
+];
